+++ framework/Furion.Extras.DatabaseAccessor.SqlSugar/Repositories/SqlSugarRepository.cs
@@ -21,11 +21,29 @@ namespace SqlSugar
         /// 构造函数
         /// </summary>
         /// <param name="serviceProvider">服务提供器</param>
-        public SqlSugarRepository(
-            IServiceProvider serviceProvider)
+        /// <param name="db"></param>
+        public SqlSugarRepository(IServiceProvider serviceProvider
+            , ISqlSugarClient db)
         {
             _serviceProvider = serviceProvider;
+            DynamicContext = Context = (SqlSugarClient)db;
+            Ado = db.Ado;
         }
+
+        /// <summary>
+        /// 数据库上下文
+        /// </summary>
+        public virtual SqlSugarClient Context { get; }
+
+        /// <summary>
+        /// 动态数据库上下文
+        /// </summary>
+        public virtual dynamic DynamicContext { get; }
+
+        /// <summary>
+        /// 原生 Ado 对象
+        /// </summary>
+        public virtual IAdo Ado { get; }
 
         /// <summary>
         /// 切换仓储
@@ -52,28 +70,21 @@ namespace SqlSugar
         private readonly ISqlSugarRepository _sqlSugarRepository;
 
         /// <summary>
-        /// 初始化 SqlSugar 客户端
-        /// </summary>
-        private readonly SqlSugarClient _db;
-
-        /// <summary>
         /// 构造函数
         /// </summary>
         /// <param name="sqlSugarRepository"></param>
-        /// <param name="db"></param>
-        public SqlSugarRepository(ISqlSugarRepository sqlSugarRepository
-            , ISqlSugarClient db)
+        public SqlSugarRepository(ISqlSugarRepository sqlSugarRepository)
         {
             _sqlSugarRepository = sqlSugarRepository;
 
-            DynamicContext = Context = _db = (SqlSugarClient)db;
-            Ado = _db.Ado;
+            DynamicContext = Context = sqlSugarRepository.Context;
+            Ado = sqlSugarRepository.Ado;
         }
 
         /// <summary>
         /// 实体集合
         /// </summary>
-        public virtual ISugarQueryable<TEntity> Entities => _db.Queryable<TEntity>();
+        public virtual ISugarQueryable<TEntity> Entities => _sqlSugarRepository.Context.Queryable<TEntity>();
 
         /// <summary>
         /// 数据库上下文
@@ -249,7 +260,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Insert(TEntity entity)
         {
-            return _db.Insertable(entity).ExecuteCommand();
+            return _sqlSugarRepository.Context.Insertable(entity).ExecuteCommand();
         }
 
         /// <summary>
@@ -259,7 +270,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Insert(params TEntity[] entities)
         {
-            return _db.Insertable(entities).ExecuteCommand();
+            return _sqlSugarRepository.Context.Insertable(entities).ExecuteCommand();
         }
 
         /// <summary>
@@ -269,7 +280,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Insert(IEnumerable<TEntity> entities)
         {
-            return _db.Insertable(entities.ToArray()).ExecuteCommand();
+            return _sqlSugarRepository.Context.Insertable(entities.ToArray()).ExecuteCommand();
         }
 
         /// <summary>
@@ -279,7 +290,7 @@ namespace SqlSugar
         /// <returns></returns>
         public int InsertReturnIdentity(TEntity insertObj)
         {
-            return _db.Insertable(insertObj).ExecuteReturnIdentity();
+            return _sqlSugarRepository.Context.Insertable(insertObj).ExecuteReturnIdentity();
         }
 
         /// <summary>
@@ -289,7 +300,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> InsertAsync(TEntity entity)
         {
-            return _db.Insertable(entity).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Insertable(entity).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -299,7 +310,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> InsertAsync(params TEntity[] entities)
         {
-            return _db.Insertable(entities).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Insertable(entities).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -309,7 +320,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> InsertAsync(IEnumerable<TEntity> entities)
         {
-            return _db.Insertable(entities.ToArray()).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Insertable(entities.ToArray()).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -319,7 +330,7 @@ namespace SqlSugar
         /// <returns></returns>
         public async Task<long> InsertReturnIdentityAsync(TEntity entity)
         {
-            return await _db.Insertable(entity).ExecuteReturnBigIdentityAsync();
+            return await _sqlSugarRepository.Context.Insertable(entity).ExecuteReturnBigIdentityAsync();
         }
 
         /// <summary>
@@ -329,7 +340,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Update(TEntity entity)
         {
-            return _db.Updateable(entity).ExecuteCommand();
+            return _sqlSugarRepository.Context.Updateable(entity).ExecuteCommand();
         }
 
         /// <summary>
@@ -339,7 +350,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Update(params TEntity[] entities)
         {
-            return _db.Updateable(entities).ExecuteCommand();
+            return _sqlSugarRepository.Context.Updateable(entities).ExecuteCommand();
         }
 
         /// <summary>
@@ -349,7 +360,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Update(IEnumerable<TEntity> entities)
         {
-            return _db.Updateable(entities.ToArray()).ExecuteCommand();
+            return _sqlSugarRepository.Context.Updateable(entities.ToArray()).ExecuteCommand();
         }
 
         /// <summary>
@@ -359,7 +370,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> UpdateAsync(TEntity entity)
         {
-            return _db.Updateable(entity).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Updateable(entity).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -369,7 +380,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> UpdateAsync(params TEntity[] entities)
         {
-            return _db.Updateable(entities).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Updateable(entities).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -379,7 +390,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> UpdateAsync(IEnumerable<TEntity> entities)
         {
-            return _db.Updateable(entities.ToArray()).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Updateable(entities.ToArray()).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -389,7 +400,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Delete(TEntity entity)
         {
-            return _db.Deleteable(entity).ExecuteCommand();
+            return _sqlSugarRepository.Context.Deleteable(entity).ExecuteCommand();
         }
 
         /// <summary>
@@ -399,7 +410,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Delete(object key)
         {
-            return _db.Deleteable<TEntity>().In(key).ExecuteCommand();
+            return _sqlSugarRepository.Context.Deleteable<TEntity>().In(key).ExecuteCommand();
         }
 
         /// <summary>
@@ -409,7 +420,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual int Delete(params object[] keys)
         {
-            return _db.Deleteable<TEntity>().In(keys).ExecuteCommand();
+            return _sqlSugarRepository.Context.Deleteable<TEntity>().In(keys).ExecuteCommand();
         }
 
         /// <summary>
@@ -419,7 +430,7 @@ namespace SqlSugar
         /// <returns></returns>
         public int Delete(Expression<Func<TEntity, bool>> whereExpression)
         {
-            return _db.Deleteable<TEntity>().Where(whereExpression).ExecuteCommand();
+            return _sqlSugarRepository.Context.Deleteable<TEntity>().Where(whereExpression).ExecuteCommand();
         }
 
         /// <summary>
@@ -429,7 +440,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> DeleteAsync(TEntity entity)
         {
-            return _db.Deleteable(entity).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Deleteable(entity).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -439,7 +450,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> DeleteAsync(object key)
         {
-            return _db.Deleteable<TEntity>().In(key).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Deleteable<TEntity>().In(key).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -449,7 +460,7 @@ namespace SqlSugar
         /// <returns></returns>
         public virtual Task<int> DeleteAsync(params object[] keys)
         {
-            return _db.Deleteable<TEntity>().In(keys).ExecuteCommandAsync();
+            return _sqlSugarRepository.Context.Deleteable<TEntity>().In(keys).ExecuteCommandAsync();
         }
 
         /// <summary>
@@ -459,7 +470,7 @@ namespace SqlSugar
         /// <returns></returns>
         public async Task<int> DeleteAsync(Expression<Func<TEntity, bool>> whereExpression)
         {
-            return await _db.Deleteable<TEntity>().Where(whereExpression).ExecuteCommandAsync();
+            return await _sqlSugarRepository.Context.Deleteable<TEntity>().Where(whereExpression).ExecuteCommandAsync();
         }
 
         /// <summary>